Rearrange weather submodule
